updates and bug fixes
[ui]: Added confirmation view
[ui]: fixed bug where ratings weren't saved
[ui]: Moved view and update to separate files
[db]: Updated schema
[db]: Changed exec to query for AddRatings
